kjhejk

+++ Restuarnt/PL/FOrder.cs
@@ -774,14 +774,20 @@ namespace Restuarnt.PL
         {
             try
             {
+                
+
                
                 if (Lable_Num.Text == "")
                 {
                       if (MessageBox.Show("هل تريد حفظ وسداد الفاتورة", "عمليه الحفظ والسداد", MessageBoxButtons.YesNo, MessageBoxIcon.Question, MessageBoxDefaultButton.Button2) == DialogResult.Yes)
                       {
 
+
                         if (gridView2.RowCount > 0)
+
                         {
+                         
+
                             if (rdb_takeaway.Checked == true)
                             {
 
@@ -1095,7 +1101,13 @@ namespace Restuarnt.PL
                     if (MessageBox.Show("هل تريد حفظ وسداد الفاتورة", "عمليه الحفظ والسداد", MessageBoxButtons.YesNo, MessageBoxIcon.Question, MessageBoxDefaultButton.Button2) == DialogResult.Yes)
                     {
 
-
+                        dt10.Clear();
+                        dt10 = o.validatePayOrder(Convert.ToInt32(Lable_Num.Text));
+                        if (dt10.Rows.Count>0)
+                        {
+                            MessageBox.Show("تم سداد الفاتورة من قبل");
+                            return;
+                        }
                         o.DeleteOrderNewRow();
 
                         int x;
